requirements.txt
appnope==0.1.2
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
bcrypt==3.2.0
black==22.8.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
ciscoconfparse==1.6.41
click==8.1.3
colorama==0.4.4
commonmark==0.9.1
cryptography==36.0.0
decorator==5.1.0
dnspython==2.1.0
dotmap==1.3.26
executing==0.8.3
future==0.18.2
idna==3.3
importlib-metadata==4.8.2
iniconfig==1.1.1
ipaddr==2.2.0
ipdb==0.13.9
ipython==8.5.0
jedi==0.18.1
Jinja2==3.1.2
junos-eznc==2.6.5
loguru==0.6.0
lxml==4.9.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
napalm==4.0.0
ncclient==0.6.13
netaddr==0.8.0
netmiko==4.1.2
netutils==1.2.0
nornir==3.3.0
nornir-jinja2==0.2.0
nornir-netmiko==0.2.0
nornir-rich @ git+https://github.com/sjhloco/nornir_rich@849923339ac0ade999059b7cd2a28890dfcaf4b2
nornir-utils==0.2.0
ntc-templates==3.1.0
orionsdk==0.3.0
packaging==21.3
paramiko==2.10.1
parso==0.8.3
passlib==1.7.4
pathspec==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.2
pluggy==1.0.0
prompt-toolkit==3.0.24
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyaml==21.10.1
pycparser==2.21
pyeapi==0.8.4
Pygments==2.11.1
PyNaCl==1.4.0
pyparsing==3.0.6
pyserial==3.5
pytest==7.1.3
PyYAML==6.0
requests==2.28.1
rich==12.6.0
ruamel.yaml==0.17.19
ruamel.yaml.clib==0.2.6
scp==0.14.1
six==1.16.0
stack-data==0.2.0
style==1.1.0
tenacity==8.0.1
textfsm==1.1.2
toml==0.10.2
tomli==2.0.1
traitlets==5.1.1
transitions==0.8.10
ttp==0.9.1
ttp-templates==0.3.1
typing_extensions==4.3.0
update==0.0.1
urllib3==1.26.12
wcwidth==0.2.5
yamlordereddictloader==0.4.0
zipp==3.7.0
